Extract refs in issue comments + refactor to GitHubAPI class
