+++ src/main/java/eu/fasten/vulnerabilityproducer/utils/PatchFarmer.java
@@ -18,20 +18,17 @@
 
 package eu.fasten.vulnerabilityproducer.utils;
 
-import com.fasterxml.jackson.databind.util.JSONPObject;
 import com.mongodb.client.FindIterable;
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 import eu.fasten.vulnerabilityproducer.db.NitriteController;
 import eu.fasten.vulnerabilityproducer.db.PatchObject;
+import eu.fasten.vulnerabilityproducer.utils.connections.GitHubAPI;
 import eu.fasten.vulnerabilityproducer.utils.connections.JavaHttpClient;
-import eu.fasten.vulnerabilityproducer.utils.connections.MongoConnector;
 import org.apache.commons.lang3.StringUtils;
 import org.bson.Document;
 import org.bson.conversions.Bson;
 import org.joda.time.DateTime;
-import org.joda.time.format.DateTimeFormat;
-import org.joda.time.format.DateTimeFormatter;
 import org.joda.time.format.ISODateTimeFormat;
 import org.jooq.tools.json.JSONArray;
 import org.jooq.tools.json.JSONObject;
@@ -60,6 +57,7 @@ public class PatchFarmer {
     public static MongoCollection<Document> mongoPullRequestsCollection;
     public static MongoCollection<Document> mongoIssuesCollection;
     public static JavaHttpClient httpClient;
+    public static GitHubAPI gitHubAPI;
     public String token;
     private final Logger logger = LoggerFactory.getLogger(PatchFarmer.class.getName());
 
@@ -73,6 +71,11 @@ public class PatchFarmer {
         mongoCommitsCollection = db.getCollection("commits");
         mongoPullRequestsCollection = db.getCollection("pull_requests");
         mongoIssuesCollection = db.getCollection("issues");
+        gitHubAPI = new GitHubAPI(httpClient, token, jsonParser);
+    }
+
+    public static GitHubAPI getGitHubAPI() {
+        return gitHubAPI;
     }
 
     /**
@@ -190,7 +193,7 @@ public class PatchFarmer {
             } else {
                 // Extra checks to get more information out of it
                 // PRs mentioning issues
-                List<String> prsMentionIssue = getPRsMentioningIssue(repoOwner, repoName, issueNumber);
+                List<String> prsMentionIssue = gitHubAPI.getPRsMentioningIssue(repoOwner, repoName, issueNumber);
                 if (prsMentionIssue.size() > 0) {
                     logger.info("Found " + prsMentionIssue.size() + " PRs mentioning issue from "
                             + repoOwner + "/" + repoName + "#" + issueNumber);
@@ -200,7 +203,7 @@ public class PatchFarmer {
                 }
 
                 // Get commit closing the issue
-                String commitsClosingIssue = getCommitClosingIssue(repoOwner, repoName, issueNumber);
+                String commitsClosingIssue = gitHubAPI.getCommitClosingIssue(repoOwner, repoName, issueNumber);
                 if (commitsClosingIssue != null) {
                     logger.info("Found " + commitsClosingIssue + " closing "
                             + repoOwner + "/" + repoName + "#" + issueNumber);
@@ -208,7 +211,7 @@ public class PatchFarmer {
                 }
 
                 // Get commits that reference issue
-                List<String> commitsReferencingIssue = getCommitsReferencingIssue(repoOwner,
+                List<String> commitsReferencingIssue = gitHubAPI.getCommitsReferencingIssue(repoOwner,
                         repoName, issueNumber);
                 if (commitsReferencingIssue.size() > 0) {
                     logger.info("Found " + commitsReferencingIssue.size() + " commits referencing issue "
@@ -217,191 +220,22 @@ public class PatchFarmer {
                 for (String commitUrl : commitsReferencingIssue) {
                     vp.addAll(parseReference(commitUrl));
                 }
+
+                // Get comments and look for references
+                List<String> potentialReferences = gitHubAPI.getReferencesInIssueComments(repoOwner,
+                        repoName, issueNumber);
+                if (potentialReferences.size() > 0) {
+                    logger.info("Found " + potentialReferences.size() + " references in comments of issue "
+                            + repoOwner + "/" + repoName + "#" + issueNumber);
+                }
+                for (String refUrl : potentialReferences) {
+                    vp.addAll(parseReference(refUrl));
+                }
             }
         } catch (Exception e) {
             e.printStackTrace();
         }
         return vp;
-    }
-
-    /**
-     * Uses API v4 to retrieve Merged Pull Requests that mention the issue under inspection
-     * @param repoOwner - Owner of the repo
-     * @param repoName - Name of the repo
-     * @param issueNumber - Issue number
-     * @return - list of String ids of those PRs
-     */
-    public List<String> getPRsMentioningIssue(String repoOwner,
-                                              String repoName,
-                                              String issueNumber) {
-        List<String> mergeCommits = new ArrayList<>();
-        var query = "query {\n" +
-                "  repository(owner: \""+ repoOwner +"\", name: \"" + repoName +"\") {\n" +
-                "    issue(number: " + issueNumber + ") {\n" +
-                "      timelineItems(first: 100) {\n" +
-                "        nodes {\n" +
-                "          ... on ClosedEvent {\n" +
-                "            closer {\n" +
-                "              ... on PullRequest {\n" +
-                "                number\n" +
-                "                title\n" +
-                "                mergeCommit {\n" +
-                "                  url\n" +
-                "                }\n" +
-                "              }\n" +
-                "            }\n" +
-                "          }\n" +
-                "        }\n" +
-                "      }\n" +
-                "    }\n" +
-                "  }\n" +
-                "}";
-        HashMap<String, String> values = new HashMap<>();
-        values.put("query", query);
-        try {
-            String jsonIssue = httpClient.sendPost("https://api.github.com/graphql", token, values);
-            // Parse the json response
-            JSONObject payload = (JSONObject) jsonParser.parse(jsonIssue);
-            JSONObject data = (JSONObject) payload.get("data");
-            if (data.get("repository") == null) return mergeCommits;
-            JSONObject repository = (JSONObject) data.get("repository");
-            if (repository.get("issue") == null)    return mergeCommits;
-            JSONArray nodes = (JSONArray) ((JSONObject) ((JSONObject) repository.get("issue")).get("timelineItems")).get("nodes");
-            for (Object nodeObj : nodes) {
-                var node = (JSONObject) nodeObj;
-                if (node.get("closer") != null) {
-                    // Grab number of the PR
-                    var closer = (JSONObject) node.get("closer");
-                    if (closer.get("mergeCommit") != null) {
-                        var mergeCommitObj = (JSONObject) closer.get("mergeCommit");
-                        if (mergeCommitObj.get("url") != null) {
-                            var mergingCommit = mergeCommitObj.get("url").toString();
-                            // Filter only commits belonging to the repo
-                            if (mergingCommit.contains(repoOwner + "/" + repoName)) {
-                                mergeCommits.add(mergingCommit);
-                            }
-                        }
-                    }
-                }
-            }
-
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return mergeCommits;
-    }
-
-    /**
-     * Looks for the commit closing a GitHub issue.
-     * @param repoOwner - Owner of the repo
-     * @param repoName - Name of the repo
-     * @param issueNumber - Issue number
-     * @return - web url link to the commit.
-     */
-    public String getCommitClosingIssue(String repoOwner,
-                                        String repoName,
-                                        String issueNumber) {
-        String commitRef = null;
-        var query = "query {\n" +
-                "  repository(owner: \""+ repoOwner +"\", name: \"" + repoName +"\") {\n" +
-                "    issue(number: " + issueNumber + ") {\n" +
-                "      timelineItems(first: 100) {\n" +
-                "        nodes {\n" +
-                "          ... on ClosedEvent {\n" +
-                "            closer {\n" +
-                "              ... on Commit {\n" +
-                "                commitUrl\n" +
-                "              }\n" +
-                "            }\n" +
-                "          }\n" +
-                "        }\n" +
-                "      }\n" +
-                "    }\n" +
-                "  }\n" +
-                "}";
-        HashMap<String, String> values = new HashMap<>();
-        values.put("query", query);
-        try {
-            String jsonIssue = httpClient.sendPost("https://api.github.com/graphql", token, values);
-            // Parse the json response
-            JSONObject payload = (JSONObject) jsonParser.parse(jsonIssue);
-            JSONObject data = (JSONObject) payload.get("data");
-            if (data.get("repository") == null) return commitRef;
-            JSONObject repository = (JSONObject) data.get("repository");
-            if (repository.get("issue") == null)    return commitRef;
-            JSONArray nodes = (JSONArray) ((JSONObject) ((JSONObject) repository.get("issue")).get("timelineItems")).get("nodes");
-            for (Object nodeObj : nodes) {
-                var node = (JSONObject) nodeObj;
-                if (node.get("closer") != null) {
-                    // Grab number of the PR
-                    var closer = (JSONObject) node.get("closer");
-                    if (closer.get("commitUrl") != null) {
-                        commitRef = ((JSONObject) node.get("closer")).get("commitUrl").toString();
-                    }
-                }
-            }
-
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return commitRef;
-    }
-
-    /**
-     * Retrieves commits referencing issues.
-     * @param repoOwner - Owner of the repo
-     * @param repoName - Name of the repo
-     * @param issueNumber - Issue number
-     * @return - list of String ids of those PRs
-     */
-    public List<String> getCommitsReferencingIssue(String repoOwner,
-                                              String repoName,
-                                              String issueNumber) {
-        List<String> commitsRef = new ArrayList<>();
-        var query = "query {\n" +
-                "  repository(owner: \""+ repoOwner +"\", name: \"" + repoName +"\") {\n" +
-                "    issue(number: " +issueNumber + ") {\n" +
-                "      timelineItems(first: 100) {\n" +
-                "        nodes {\n" +
-                "          ... on ReferencedEvent {\n" +
-                "            commit {\n" +
-                "              commitUrl\n" +
-                "            }\n" +
-                "          }\n" +
-                "        }\n" +
-                "      }\n" +
-                "    }\n" +
-                "  }\n" +
-                "}";
-        HashMap<String, String> values = new HashMap<>();
-        values.put("query", query);
-        try {
-            String jsonIssue = httpClient.sendPost("https://api.github.com/graphql", token, values);
-            // Parse the json response
-            JSONObject payload = (JSONObject) jsonParser.parse(jsonIssue);
-            JSONObject data = (JSONObject) payload.get("data");
-            if (data.get("repository") == null) return commitsRef;
-            JSONObject repository = (JSONObject) data.get("repository");
-            if (repository.get("issue") == null)    return commitsRef;
-            JSONArray nodes = (JSONArray) ((JSONObject) ((JSONObject) repository.get("issue")).get("timelineItems")).get("nodes");
-            for (Object nodeObj : nodes) {
-                var node = (JSONObject) nodeObj;
-                if (node.get("commit") != null) {
-                    // Grab number of the PR
-                    var closer = (JSONObject) node.get("commit");
-                    if (closer.get("commitUrl") != null) {
-                        var commitUrl = ((JSONObject) node.get("commit")).get("commitUrl").toString();
-                        if (commitUrl.contains("/" + repoOwner + "/" + repoName + "/")) {
-                            commitsRef.add(commitUrl);
-                        }
-                    }
-                }
-            }
-
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return commitsRef;
     }
 
     /**
@@ -797,7 +631,7 @@ public class PatchFarmer {
      * @param text
      * @return
      */
-    private List<String> findPatchLinks(String text) {
+    public static List<String> findPatchLinks(String text) {
         List<String> potentialPatches = new ArrayList<>();
         List<Matcher> matchers = new ArrayList<>();
