+++ src/test/java/eu/fasten/vulnerabilityproducer/PatchFarmerTest.java
@@ -652,6 +652,11 @@ public class PatchFarmerTest {
     }
 
     @Test
+    public void testGetReferencesInIssueComments() {
+        //TODO
+    }
+
+    @Test
     public void testGetCommitThatClosedIssue() throws Exception {
         var ref = "https://github.com/hapijs/subtext/issues/72";
         var issueJson = readFile("./src/test/resources/patches/commitClosesIssue/issueAPIv4.json");
